consolidate LogFieldSetup, TreeFieldSetup, Log, LogError
