+++ src/FScruiser.Xamarin/ViewModels/SaleSelectViewModel.cs
@@ -1,8 +1,6 @@
 ﻿using FScruiser.XF.Services;
 using NatCruise.Data;
-using NatCruise.Data.Abstractions;
 using NatCruise.Models;
-using NatCruise.Navigation;
 using Prism.Common;
 using System;
 using System.Collections.Generic;
