+++ src/NatCruise.Core/Data/Abstractions/IFieldSetupDataservice.cs
@@ -1,20 +1,29 @@
-﻿using NatCruise.Data;
-using NatCruise.Design.Models;
+﻿using NatCruise.Models;
 using System.Collections.Generic;
 
-namespace NatCruise.Design.Data
+namespace NatCruise.Data
 {
     public interface IFieldSetupDataservice : IDataservice
     {
+        #region TreeFieldSetups
+
         IEnumerable<TreeFieldSetup> GetTreeFieldSetups(string stratumCode);
+
         IEnumerable<TreeFieldSetup> GetTreeFieldSetups(string stratumCode, string sampleGroupCode);
+
         void UpsertTreeFieldSetup(TreeFieldSetup tfs);
+
         void DeleteTreeFieldSetup(TreeFieldSetup tfs);
 
         //void SetTreeFieldsFromStratumDefault(string stratumCode, StratumDefault sd);
 
         void SetTreeFieldsFromStratumTemplate(string stratumCode, string stratumTemplateName);
 
+        #endregion TreeFieldSetups
+
+        #region LogFieldSetup
+
+        IEnumerable<LogFieldSetup> GetLogFieldSetupsByTreeID(string tree_guid);
 
         IEnumerable<LogFieldSetup> GetLogFieldSetups(string stratumCode);
 
@@ -23,5 +32,7 @@ namespace NatCruise.Design.Data
         void DeleteLogFieldSetup(LogFieldSetup lfs);
 
         void SetLogFieldsFromStratumTemplate(string stratumCode, string stratumTemplateName);
+
+        #endregion LogFieldSetup
     }
 }
