+++ src/NatCruise.Core/Models/LogError.cs
@@ -1,0 +1,11 @@
+﻿using Prism.Mvvm;
+
+namespace NatCruise.Models
+{
+    public class LogError : BindableBase
+    {
+        public string LogID { get; set; }
+
+        public int LogNumber { get; set; }
+    }
+}
