+++ src/NatCruise.Cruise/Data/DataserviceProvider.cs
@@ -2,7 +2,6 @@
 using NatCruise.Core.Services;
 using NatCruise.Cruise.Services;
 using NatCruise.Data;
-using NatCruise.Data.Abstractions;
 using Prism.Ioc;
 using System;
 
@@ -35,6 +34,12 @@ namespace NatCruise.Cruise.Data
 
         public override IDataservice GetDataservice(Type type)
         {
+            var ds = base.GetDataservice(type);
+            if (ds != null)
+            {
+                return ds;
+            }
+
             var cruiseID = CruiseID;
             var database = Database;
             var deviceID = DeviceID;
@@ -45,17 +50,8 @@ namespace NatCruise.Cruise.Data
             if (cruiseID == null)
             { throw new InvalidOperationException("DataserviceProvider: no cruise selected"); }
 
-            if (typeof(ICuttingUnitDataservice).IsAssignableFrom(type))
-            { return new CuttingUnitDataservice(database, cruiseID, deviceID); }
-
             if (typeof(IPlotDataservice).IsAssignableFrom(type))
             { return new PlotDataservice(database, cruiseID, deviceID); }
-
-            if (typeof(ITreeDataservice).IsAssignableFrom(type))
-            { return new TreeDataservice(database, cruiseID, deviceID); }
-
-            if (typeof(ILogDataservice).IsAssignableFrom(type))
-            { return new LogDataservice(database, cruiseID, deviceID); }
 
             if (typeof(IPlotTreeDataservice).IsAssignableFrom(type))
             { return new PlotTreeDataservice(database, cruiseID, deviceID, GetDataservice<ISampleInfoDataservice>()); }
@@ -69,47 +65,17 @@ namespace NatCruise.Cruise.Data
             if (typeof(ITallyDataservice).IsAssignableFrom(type))
             { return new TallyDataservice(database, cruiseID, deviceID, GetDataservice<ISampleInfoDataservice>()); }
 
-            if(typeof(ISampleInfoDataservice).IsAssignableFrom(type))
-            {
-                return new SamplerInfoDataservice(database, cruiseID, deviceID);
-            }
-
-            if(typeof(ITreeFieldDataservice).IsAssignableFrom(type))
-            { return new TreeFieldDataservice(database, cruiseID, deviceID); }
+            if (typeof(ISampleInfoDataservice).IsAssignableFrom(type))
+            { return new SamplerInfoDataservice(database, cruiseID, deviceID); }
 
             if (typeof(ITallyPopulationDataservice).IsAssignableFrom(type))
             { return new TallyPopulationDataservice(database, cruiseID, deviceID); }
 
             if (typeof(ISampleInfoDataservice).IsAssignableFrom(type))
             { return new SamplerInfoDataservice(database, cruiseID, deviceID); }
-
-            if (typeof(IStratumDataservice).IsAssignableFrom(type))
-            {
-                return new StratumDataservice(database, cruiseID, deviceID);
-            }
-            if (typeof(ISampleGroupDataservice).IsAssignableFrom(type))
-            {
-                return new SampleGroupDataservice(database, cruiseID, deviceID);
-            }
-            if (typeof(ISubpopulationDataservice).IsAssignableFrom(type))
-            {
-                return new SubpopulationDataservice(database, cruiseID, deviceID);
-            }
-            if (typeof(ITreeErrorDataservice).IsAssignableFrom(type))
-            {
-                return new TreeErrorDataservice(database, cruiseID, deviceID);
-            }
-            if (typeof(ITreeFieldValueDataservice).IsAssignableFrom(type))
-            {
-                return new TreeFieldValueDataservice(database, cruiseID, deviceID);
-            }
-
             else
             {
-
                 throw new InvalidOperationException("no dataservice found for type " + type.FullName);
-
-                
 
                 //return null;
             }
@@ -117,23 +83,15 @@ namespace NatCruise.Cruise.Data
 
         public override void RegisterDataservices(IContainerRegistry containerRegistry)
         {
-            containerRegistry.Register<ICuttingUnitDataservice>(x => GetDataservice<ICuttingUnitDataservice>());
-            containerRegistry.Register<IStratumDataservice>(x => GetDataservice<IStratumDataservice>());
-            containerRegistry.Register<ISampleGroupDataservice>(x => GetDataservice<ISampleGroupDataservice>());
-            containerRegistry.Register<ISubpopulationDataservice>(x => GetDataservice<ISubpopulationDataservice>());
+            base.RegisterDataservices(containerRegistry);
+
             containerRegistry.Register<IPlotDataservice>(x => GetDataservice<IPlotDataservice>());
-            containerRegistry.Register<ITreeDataservice>(x => GetDataservice<ITreeDataservice>());
-            containerRegistry.Register<ITreeErrorDataservice>(x => GetDataservice<ITreeErrorDataservice>());
-            containerRegistry.Register<ITreeFieldValueDataservice>(x => GetDataservice<ITreeFieldValueDataservice>());
-            containerRegistry.Register<ILogDataservice>(x => GetDataservice<ILogDataservice>());
+            containerRegistry.Register<IPlotTreeDataservice>(e => GetDataservice<IPlotTreeDataservice>());
             containerRegistry.Register<ISampleSelectorDataService>(x => GetDataservice<ISampleSelectorDataService>());
-            containerRegistry.Register<ISaleDataservice>(x => GetDataservice<ISaleDataservice>());
             containerRegistry.Register<IFixCNTDataservice>(x => GetDataservice<IFixCNTDataservice>());
             containerRegistry.Register<ITallyDataservice>(x => GetDataservice<ITallyDataservice>());
-            containerRegistry.Register<IPlotTreeDataservice>(e => GetDataservice<IPlotTreeDataservice>());
             containerRegistry.Register<ITallyPopulationDataservice>(x => GetDataservice<ITallyPopulationDataservice>());
             containerRegistry.Register<ISampleInfoDataservice>(x => GetDataservice<ISampleInfoDataservice>());
-            containerRegistry.Register<ITreeFieldDataservice>(x => GetDataservice<ITreeFieldDataservice>());
         }
     }
 }
