--- src/NatCruise.Cruise/Models/LogFieldSetup.cs
@@ -1,12 +1,0 @@
-﻿using FMSC.ORM.EntityModel.Attributes;
-
-namespace NatCruise.Cruise.Models
-{
-    [Table("LogFieldSetup")]
-    public class LogFieldSetup
-    {
-        public string Field { get; set; }
-
-        public string Heading { get; set; }
-    }
-}
