+++ src/NatCruise.Design/Validation/TreeFieldSetupValidator.cs
@@ -1,10 +1,5 @@
 ﻿using FluentValidation;
-using NatCruise.Design.Models;
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
+using NatCruise.Models;
 
 namespace NatCruise.Design.Validation
 {
@@ -21,7 +16,6 @@ namespace NatCruise.Design.Validation
                 .GreaterThanOrEqualTo(0.0).When(x => x.DefaultValueReal != null)
                 .WithSeverity(Severity.Error)
                 .WithName("Default Value");
-
         }
     }
 }
