+++ src/NatCruise.Design/ViewModels/StratumLogFieldSetupViewModel.cs
@@ -1,4 +1,5 @@
-﻿using NatCruise.Design.Data;
+﻿using NatCruise.Data;
+using NatCruise.Design.Data;
 using NatCruise.Design.Models;
 using NatCruise.Models;
 using Prism.Commands;
